Rename instructlab.config to instructlab.configuration
This changed in the instructlab repo, so match it.
