+++ requirements.txt
@@ -6,4 +6,4 @@ openai>=1.13.3,<2.0.0
 rouge_score
 tqdm>=4.66.2,<5.0.0
 # TODO: remove 'instructlab' once https://github.com/instructlab/sdg/issues/6 is resolved
-instructlab
+instructlab>=0.17.0
